+++ CHANGES.rst
@@ -4,6 +4,12 @@ Changelog
 
 All major versions including are listed below.
 
+Version 0.1.0.3b1
+~~~~~~~~~~~~~~~~~
+
+- Add function to prepare system configurations for DS18B20 sensors
+- Bug fixes in eems console script
+
 Version 0.1.0.2b1
 ~~~~~~~~~~~~~~~~~
 
@@ -13,18 +19,15 @@ Version 0.1.0.2b1
 - Updated csv export file structure
 - Moved csv handling to separated module
 
-
 Version 0.1.0.1b1
 ~~~~~~~~~~~~~~~~~
 
 Open beta for testing issues.
 
-
 Version 0.0.4.3a1
 ~~~~~~~~~~~~~~~~~
 
 Closed alpha version.
-
 
 Version 0.0.3.dev1
 ~~~~~~~~~~~~~~~~~~